requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.0
numpy>=1.20.0
pdfplumber>=0.10.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.0.0
